Refactor minimizer
